ULTIMA: Split up the Debugger classes
