+++ engines/ultima/ultima4/core/debugger.cpp
@@ -48,7 +48,90 @@ namespace Ultima4 {
 
 Debugger *g_debugger;
 
-Debugger::Debugger() : Shared::Debugger() {
+static bool strToBool(const char *s) {
+	return s && tolower(*s) == 't';
+}
+
+static int strToInt(const char *s) {
+	if (!*s)
+		// No string at all
+		return 0;
+	else if (toupper(s[strlen(s) - 1]) != 'H')
+		// Standard decimal string
+		return atoi(s);
+
+	// Hexadecimal string
+	uint tmp = 0;
+	int read = sscanf(s, "%xh", &tmp);
+	if (read < 1)
+		error("strToInt failed on string \"%s\"", s);
+	return (int)tmp;
+}
+
+static void splitString(const Common::String &str,
+		Common::StringArray &argv) {
+	// Clear the vector
+	argv.clear();
+
+	bool quoted = false;
+	Common::String::const_iterator it;
+	int ch;
+	Common::String arg;
+
+	for (it = str.begin(); it != str.end(); ++it) {
+		ch = *it;
+
+		// Toggle quoted string handling
+		if (ch == '\"') {
+			quoted = !quoted;
+			continue;
+		}
+
+		// Handle \\, \", \', \n, \r, \t
+		if (ch == '\\') {
+			Common::String::const_iterator next = it + 1;
+			if (next != str.end()) {
+				if (*next == '\\' || *next == '\"' || *next == '\'') {
+					ch = *next;
+					++it;
+				} else if (*next == 'n') {
+					ch = '\n';
+					++it;
+				} else if (*next == 'r') {
+					ch = '\r';
+					++it;
+				} else if (*next == 't') {
+					ch = '\t';
+					++it;
+				} else if (*next == ' ') {
+					ch = ' ';
+					++it;
+				}
+			}
+		}
+
+		// A space, a tab, line feed, carriage return
+		if (!quoted && (ch == ' ' || ch == '\t' || ch == '\n' || ch == '\r')) {
+			// If we are not empty then we are at the end of the arg
+			// otherwise we will ignore the extra chars
+			if (!arg.empty()) {
+				argv.push_back(arg);
+				arg.clear();
+			}
+
+			continue;
+		}
+
+		// Add the charater to the string
+		arg += ch;
+	}
+
+	// Push any arg if it's left
+	if (!arg.empty())
+		argv.push_back(arg);
+}
+
+Debugger::Debugger() : GUI::Debugger() {
 	g_debugger = this;
 	_collisionOverride = false;
 	_disableCombat = false;
@@ -194,7 +277,7 @@ bool Debugger::handleCommand(int argc, const char **argv, bool &keepRunning) {
 		}
 	}
 
-	bool result = Shared::Debugger::handleCommand(argc, argv, keepRunning);
+	bool result = GUI::Debugger::handleCommand(argc, argv, keepRunning);
 
 	if (result) {
 		Controller *ctl = eventHandler->getController();
@@ -1885,6 +1968,34 @@ bool Debugger::cmdListTriggers(int argc, const char **argv) {
 	}
 
 	return isDebuggerActive();
+}
+
+void Debugger::executeCommand(const Common::String &cmd) {
+	// Split up the command, and form a const char * array
+	Common::StringArray args;
+	splitString(cmd, args);
+
+	Common::Array<const char *> argv;
+	for (uint idx = 0; idx < args.size(); ++idx)
+		argv.push_back(args[idx].c_str());
+
+	// Execute the command
+	executeCommand(argv.size(), &argv[0]);
+}
+
+void Debugger::executeCommand(int argc, const char **argv) {
+	if (argc <= 0)
+		return;
+
+	bool keepRunning = false;
+	if (!handleCommand(argc, argv, keepRunning)) {
+		debugPrintf("Unknown command - %s\n", argv[0]);
+		keepRunning = true;
+	}
+
+	// If any message occurred, then we need to ensure the debugger is opened if it isn't already
+	if (keepRunning)
+		attach();
 }
 
 } // End of namespace Ultima4
